Move message_payload.h into content/public
