+++ content/browser/BUILD.gn
@@ -2938,7 +2938,6 @@ source_set("browser") {
       "android/javascript_injector.h",
       "android/load_url_params.cc",
       "android/message_payload.cc",
-      "android/message_payload.h",
       "android/navigation_handle_proxy.cc",
       "android/navigation_handle_proxy.h",
       "android/nfc_host.cc",
