+++ content/browser/android/message_payload.cc
@@ -1,7 +1,7 @@
 // Copyright 2022 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-#include "content/browser/android/message_payload.h"
+#include "content/public/browser/android/message_payload.h"
 
 #include <string>
 #include <utility>
