+++ content/browser/android/message_payload_unittest.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "content/browser/android/message_payload.h"
+#include "content/public/browser/android/message_payload.h"
 #include <cstddef>
 #include <string>
