+++ content/public/browser/BUILD.gn
@@ -568,6 +568,7 @@ source_set("browser_sources") {
       "android/java_interfaces.h",
       "android/media_url_interceptor_register.cc",
       "android/media_url_interceptor_register.h",
+      "android/message_payload.h",
       "android/message_payload_type.h",
       "android/message_port_helper.h",
       "android/motion_event_action.h",
