+++ content/public/browser/android/message_payload.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#ifndef CONTENT_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
-#define CONTENT_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
+#ifndef CONTENT_PUBLIC_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
+#define CONTENT_PUBLIC_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
 
 #include "base/android/jni_android.h"
 #include "base/android/scoped_java_ref.h"
@@ -29,4 +29,4 @@ CONTENT_EXPORT blink::WebMessagePayload ConvertToWebMessagePayloadFromJava(
 
 }  // namespace content::android
 
-#endif  // CONTENT_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
+#endif  // CONTENT_PUBLIC_BROWSER_ANDROID_MESSAGE_PAYLOAD_H_
